feat: advanced agrouped endpoints

+++ DoubleVPartnersBackEnd/DoubleVPartners.BackEnd.Domain/Common/Contracts/Persistence/IGenericRepository.cs
@@ -12,7 +12,7 @@ namespace DoubleVPartners.BackEnd.Domain.Common.Contracts.Persistence
         Task<IEnumerable<T>> Where(Expression<Func<T, bool>> predicate, string includes = "", PaginationFilter? pagination = null, OrderByClausure<T>? orderBy = null);
         Task<int> Count();
         Task<int> Count(Expression<Func<T, bool>> predicate);
-        Task<double> Sum(Expression<Func<T, double>> summaryPredicate, Expression<Func<T, bool>>? filterPredicate = null);
+        Task<decimal> Sum(Expression<Func<T, decimal>> summaryPredicate, Expression<Func<T, bool>>? filterPredicate = null);
         void Update(T entity);
         void UpdateRange(IEnumerable<T> entity);
         void Add(T entity);
